rollinggo-hotel-uv/pyproject.toml
[project]
name = "rollinggo-mcp"
version = "0.1.0"
description = "RollingGo hotel MCP server for stdio clients."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastmcp==3.1.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
]

[project.scripts]
rollinggo-mcp = "rollinggo_mcp.__main__:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rollinggo_mcp"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
